FJA-43
Tech task: Change LegalEntityId to long/biginteger

+++ src/SFA.DAS.Rofjaa.Application/Agencies/Queries/GetAgency/GetAgencyQuery.cs
@@ -4,6 +4,6 @@ namespace SFA.DAS.Rofjaa.Application.Agencies.Queries.GetAgency
 {
     public class GetAgencyQuery : IRequest<GetAgencyResult>
     {
-        public int LegalEntityId { get ; set ; }
+        public long LegalEntityId { get ; set ; }
     }
 }
