+++ src/SFA.DAS.Rofjaa.Application/Agencies/Queries/GetAgency/GetAgencyResult.cs
@@ -4,7 +4,7 @@ namespace SFA.DAS.Rofjaa.Application.Agencies.Queries.GetAgency
 {
     public class GetAgencyResult
     {
-        public int LegalEntityId { get; set; }
+        public long LegalEntityId { get; set; }
         public bool IsGrantFunded { get; set; }
     }
 }
