+++ src/SFA.DAS.Rofjaa.Database/Tables/Agency.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Agency]
 (
-	[LegalEntityId] INT NOT NULL PRIMARY KEY,
+	[LegalEntityId] bigint NOT NULL PRIMARY KEY,
 	[IsGrantFunded] bit NOT NULL DEFAULT 0
 )
 GO
