+++ src/SFA.DAS.Rofjaa.Domain/Types/Agency.cs
@@ -7,7 +7,7 @@ namespace SFA.DAS.Rofjaa.Domain.Types
     public class Agency
     {
         [JsonProperty("LegalEntityId")]
-        public int LegalEntityId { get; set; }
+        public long LegalEntityId { get; set; }
 
         [JsonProperty("IsGrantFunded")]
         public bool IsGrantFunded { get; set; }
